Adds collection classes to differentiate collection methods
